Complete project modularization - Refactored code across the entire project. Some corrections pending.
