+++ app/src/main/java/com/example/handappmobile_epn/bt/BluetoothConstants.kt
@@ -1,0 +1,3 @@
+package com.example.handappmobile_epn.bt
+
+const val REQUEST_ENABLE_BT = 1
